[gnmi]: Move SHOW CLI helpers to common
