+++ tests/common/helpers/gnmi_cli_helpers.py
